requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.0
Brotli==1.1.0
certifi==2024.8.30
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
deprecation==2.1.0
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
frozenlist==1.4.1
gotrue==2.8.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
kombu==5.4.2
msgpack==1.1.0
multidict==6.1.0
packaging==24.1
pillow==10.4.0
postgrest==0.16.11
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
realtime==2.0.2
redis==5.0.8
requests==2.32.3
schedule==1.2.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
whitenoise==6.7.0
yarl==1.11.1
